// src/app/interfaces/producto-request.ts
export interface ProductoRequest {
  codigo: string;
  nombre: string;
  descripcion: string;
  largo: number;
  ancho: number;
  profundidad: number;
  stock: number;
  presentacion: string;
  rendimiento: number;
  textura: string;
  sugerencias: string;
  rucFilial: string;
  rucProveedor: string;
  idCategoria: number;
}